`chore: Update URLs for allauth and trainers app`
The URLs in the `digidex` and `trainers` apps have been updated. The `allauth` URLs have been changed to include the `/accounts` prefix, and the trainers app URLs have been modified to use the `trainers` prefix. This change ensures that the URLs are more descriptive and aligned with the app names.